backend/requirements.txt
asgiref==3.7.2
boto3==1.28.53
botocore==1.31.53
Django==4.2
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
jmespath==1.0.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==1.26.16